stm32/usb: merge usb and usb_otg into single module.
